Add slide indicator
Add a dot (&#9673;) for every slide at the top as an indication where
we are in the tour. Dots are links to the oter slides with tool tips
with their names.

All dots are white, only the current slide should be black.

Only 1-start.html has been treated with this so far.